Rename /registration to /login
